// sourcecode/BTrak.Packages/workflows/src/lib/workflows/models/workflow-model.ts
export class WorkflowModel {
    id:string;
    formtypeId: string;
    formtypeName: string;
    formName: string;
    workFlowTypeId: string;
    workflowName: string;
    isStatus: boolean;
    description: string;
    action:string;
    fieldname:any;
    fieldNames:any;
    dateofexecution:any;
    triggerMonth:any;
    triggerDay:any;
    triggerStartList:any;
    triggerStartDays:any;
    triggerEndList:any;
    triggerEndDays:any;
    timeofexecution:any;
    executionoccurrence:any;
    timezone:any;
    dataJson:string;
    xml:any;
    cronExpression: string;
    cronRadio:string;
    workflowItems: any;
    workflowItemsByOrder: any;
    isArchived:boolean;
    criterialSteps: any;
    companyModuleId: string;
    formId: string;
    dataSourceId : string;
    tasks: any[];
    fieldUniqueId: string;
    dateFieldName: string;
    offsetMinutes : any;
    selectedTab: any;
    formIds : any;
}